Percent From Fraction
Earlier, when creating a `PercentValue` from a fraction (ie: `PercentValue::fromFraction()` if the calculation resulted in an integer, the percent value object could not be created as a float is required. Now, we cast the value passed to the parent to a float to ensure the object can be created.