Float Supports Whole
A `FloatValue` will support an `int` and automatically cast the raw value to a float. This helps when creating a `FloatValue` from a dynamic value where said value may be an `int`.

Added `PercentValue::fromWhole()` to create a `PercentValue` from a whole number. This makes it easier to work with `PercentValue` when the data source provides percents as numbers between zero and 100.